requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
google-genai==0.2.2
Pillow==11.0.0
